revise: moves `check` to top level subcommand
